Remove dependency away from device.isOnline
**This commit is Grove specific**

device.isOnline is questionably reliable. This commit moves away from that and instead calls a blank function on the device itself.

Calls and waits for a timeout up to three times before reporting the device connection as a failure.

Still needs proper QA testing (both this method and spark's device.isOnline)